+++ hydra/core/state_migrations.py
@@ -17,6 +17,8 @@ migrate_v12_to_v13 = calls_vk.migrate_v12_to_v13
 migrate_v13_to_v14 = calls_vk.migrate_v13_to_v14
 migrate_v14_to_v15 = calls_vk.migrate_v14_to_v15
 migrate_v15_to_v16 = calls_vk.migrate_v15_to_v16
+migrate_v16_to_v17 = calls_vk.migrate_v16_to_v17
+migrate_v17_to_v18 = calls_vk.migrate_v17_to_v18
 def migrate_v0_to_v1(data: dict) -> dict:
     data["version"] = 1
     data.setdefault("install", {})
@@ -110,6 +112,8 @@ MIGRATIONS: dict[int, Migration] = {
     13: calls_vk.migrate_v13_to_v14,
     14: calls_vk.migrate_v14_to_v15,
     15: calls_vk.migrate_v15_to_v16,
+    16: calls_vk.migrate_v16_to_v17,
+    17: calls_vk.migrate_v17_to_v18,
 }